refactor(editor): clarify point deletion steps
